+++ persistent-sqlite/Database/Persist/Sqlite.hs
@@ -50,7 +50,7 @@ import Control.Concurrent (threadDelay)
 import qualified Control.Exception as E
 import Control.Monad (forM_)
 import Control.Monad.IO.Unlift (MonadIO (..), MonadUnliftIO, askRunInIO, withRunInIO, withUnliftIO, unliftIO, withRunInIO)
-import Control.Monad.Logger (NoLoggingT, runNoLoggingT, MonadLogger, logWarn, runLoggingT)
+import Control.Monad.Logger (NoLoggingT, runNoLoggingT, MonadLoggerIO, logWarn, runLoggingT, askLoggerIO)
 import Control.Monad.Trans.Reader (ReaderT, runReaderT, withReaderT)
 import Control.Monad.Trans.Writer (runWriterT)
 import Data.Acquire (Acquire, mkAcquire, with)
@@ -81,7 +81,7 @@ import qualified Database.Sqlite as Sqlite
 -- Note that this should not be used with the @:memory:@ connection string, as
 -- the pool will regularly remove connections, destroying your database.
 -- Instead, use 'withSqliteConn'.
-createSqlitePool :: (MonadLogger m, MonadUnliftIO m)
+createSqlitePool :: (MonadLoggerIO m, MonadUnliftIO m)
                  => Text -> Int -> m (Pool SqlBackend)
 createSqlitePool = createSqlitePoolFromInfo . conStringToInfo
 
@@ -92,14 +92,14 @@ createSqlitePool = createSqlitePoolFromInfo . conStringToInfo
 -- Instead, use 'withSqliteConn'.
 --
 -- @since 2.6.2
-createSqlitePoolFromInfo :: (MonadLogger m, MonadUnliftIO m)
+createSqlitePoolFromInfo :: (MonadLoggerIO m, MonadUnliftIO m)
                          => SqliteConnectionInfo -> Int -> m (Pool SqlBackend)
 createSqlitePoolFromInfo connInfo = createSqlPool $ openWith const connInfo
 
 -- | Run the given action with a connection pool.
 --
 -- Like 'createSqlitePool', this should not be used with @:memory:@.
-withSqlitePool :: (MonadUnliftIO m, MonadLogger m)
+withSqlitePool :: (MonadUnliftIO m, MonadLoggerIO m)
                => Text
                -> Int -- ^ number of connections to open
                -> (Pool SqlBackend -> m a) -> m a
@@ -110,18 +110,18 @@ withSqlitePool connInfo = withSqlPool . openWith const $ conStringToInfo connInf
 -- Like 'createSqlitePool', this should not be used with @:memory:@.
 --
 -- @since 2.6.2
-withSqlitePoolInfo :: (MonadUnliftIO m, MonadLogger m)
+withSqlitePoolInfo :: (MonadUnliftIO m, MonadLoggerIO m)
                    => SqliteConnectionInfo
                    -> Int -- ^ number of connections to open
                    -> (Pool SqlBackend -> m a) -> m a
 withSqlitePoolInfo connInfo = withSqlPool $ openWith const connInfo
 
-withSqliteConn :: (MonadUnliftIO m, MonadLogger m)
+withSqliteConn :: (MonadUnliftIO m, MonadLoggerIO m)
                => Text -> (SqlBackend -> m a) -> m a
 withSqliteConn = withSqliteConnInfo . conStringToInfo
 
 -- | @since 2.6.2
-withSqliteConnInfo :: (MonadUnliftIO m, MonadLogger m)
+withSqliteConnInfo :: (MonadUnliftIO m, MonadLoggerIO m)
                    => SqliteConnectionInfo -> (SqlBackend -> m a) -> m a
 withSqliteConnInfo = withSqlConn . openWith const
 
@@ -184,7 +184,7 @@ wrapConnection = wrapConnectionInfo (mkSqliteConnectionInfo "")
 -- | Retry if a Busy is thrown, following an exponential backoff strategy.
 --
 -- @since 2.9.3
-retryOnBusy :: (MonadUnliftIO m, MonadLogger m) => m a -> m a
+retryOnBusy :: (MonadUnliftIO m, MonadLoggerIO m) => m a -> m a
 retryOnBusy action =
   start $ take 20 $ delays 1000
   where
@@ -210,7 +210,7 @@ retryOnBusy action =
 --
 -- @since 2.9.3
 waitForDatabase
-    :: (MonadUnliftIO m, MonadLogger m, BackendCompatible SqlBackend backend)
+    :: (MonadUnliftIO m, MonadLoggerIO m, BackendCompatible SqlBackend backend)
     => ReaderT backend m ()
 waitForDatabase = retryOnBusy $ rawExecute "SELECT 42" []
 
@@ -261,7 +261,9 @@ wrapConnectionInfo connInfo conn logFunc = do
         , connBegin = \f _ -> helper "BEGIN" f
         , connCommit = helper "COMMIT"
         , connRollback = ignoreExceptions . helper "ROLLBACK"
-        , connEscapeName = escape
+        , connEscapeFieldName = escape . unFieldNameDB
+        , connEscapeTableName = escape . unEntityNameDB . entityDB
+        , connEscapeRawName = escape
         , connNoLimit = "LIMIT -1"
         , connRDBMS = "sqlite"
         , connLimitOffset = decorateSQLWithLimitOffset "LIMIT -1"
@@ -321,9 +323,9 @@ insertSql' ent vals =
           ISRManyKeys sql vals
             where sql = T.concat
                     [ "INSERT INTO "
-                    , escape $ entityDB ent
+                    , escapeE $ entityDB ent
                     , "("
-                    , T.intercalate "," $ map (escape . fieldDB) cols
+                    , T.intercalate "," $ map (escapeF . fieldDB) cols
                     , ") VALUES("
                     , T.intercalate "," (map (const "?") cols)
                     , ")"
@@ -333,19 +335,19 @@ insertSql' ent vals =
             where
               sel = T.concat
                   [ "SELECT "
-                  , escape $ fieldDB (entityId ent)
+                  , escapeF $ fieldDB (entityId ent)
                   , " FROM "
-                  , escape $ entityDB ent
+                  , escapeE $ entityDB ent
                   , " WHERE _ROWID_=last_insert_rowid()"
                   ]
               ins = T.concat
                   [ "INSERT INTO "
-                  , escape $ entityDB ent
+                  , escapeE $ entityDB ent
                   , if null cols
                         then " VALUES(null)"
                         else T.concat
                           [ "("
-                          , T.intercalate "," $ map (escape . fieldDB) $ cols
+                          , T.intercalate "," $ map (escapeF . fieldDB) $ cols
                           , ") VALUES("
                           , T.intercalate "," (map (const "?") cols)
                           , ")"
@@ -409,7 +411,7 @@ migrate' allDefs getter val = do
     let (cols, uniqs, fdefs) = sqliteMkColumns allDefs val
     let newSql = mkCreateTable False def (filter (not . safeToRemove val . cName) cols, uniqs, fdefs)
     stmt <- getter "SELECT sql FROM sqlite_master WHERE type='table' AND name=?"
-    oldSql' <- with (stmtQuery stmt [PersistText $ unDBName table])
+    oldSql' <- with (stmtQuery stmt [PersistText $ unEntityNameDB table])
       (\src -> runConduit $ src .| go)
     case oldSql' of
         Nothing -> return $ Right [(False, newSql)]
@@ -451,7 +453,9 @@ mockMigration mig = do
                    , connBegin = \f _ -> helper "BEGIN" f
                    , connCommit = helper "COMMIT"
                    , connRollback = ignoreExceptions . helper "ROLLBACK"
-                   , connEscapeName = escape
+                   , connEscapeFieldName = escape . unFieldNameDB
+                   , connEscapeTableName = escape . unEntityNameDB . entityDB
+                   , connEscapeRawName = escape
                    , connNoLimit = "LIMIT -1"
                    , connRDBMS = "sqlite"
                    , connLimitOffset = decorateSQLWithLimitOffset "LIMIT -1"
@@ -473,10 +477,10 @@ mockMigration mig = do
 
 -- | Check if a column name is listed as the "safe to remove" in the entity
 -- list.
-safeToRemove :: EntityDef -> DBName -> Bool
-safeToRemove def (DBName colName)
+safeToRemove :: EntityDef -> FieldNameDB -> Bool
+safeToRemove def (FieldNameDB colName)
     = any (elem FieldAttrSafeToRemove . fieldAttrs)
-    $ filter ((== DBName colName) . fieldDB)
+    $ filter ((== FieldNameDB colName) . fieldDB)
     $ entityFields def
 
 getCopyTable :: [EntityDef]
@@ -484,9 +488,9 @@ getCopyTable :: [EntityDef]
              -> EntityDef
              -> IO [(Bool, Text)]
 getCopyTable allDefs getter def = do
-    stmt <- getter $ T.concat [ "PRAGMA table_info(", escape table, ")" ]
+    stmt <- getter $ T.concat [ "PRAGMA table_info(", escapeE table, ")" ]
     oldCols' <- with (stmtQuery stmt []) (\src -> runConduit $ src .| getCols)
-    let oldCols = map DBName $ filter (/= "id") oldCols' -- need to update for table id attribute ?
+    let oldCols = map FieldNameDB $ filter (/= "id") oldCols' -- need to update for table id attribute ?
     let newCols = filter (not . safeToRemove def) $ map cName cols
     let common = filter (`elem` oldCols) newCols
     return [ (False, tmpSql)
@@ -506,30 +510,30 @@ getCopyTable allDefs getter def = do
                 return $ name : names
             Just y -> error $ "Invalid result from PRAGMA table_info: " ++ show y
     table = entityDB def
-    tableTmp = DBName $ unDBName table <> "_backup"
+    tableTmp = EntityNameDB $ unEntityNameDB table <> "_backup"
     (cols, uniqs, fdef) = sqliteMkColumns allDefs def
     cols' = filter (not . safeToRemove def . cName) cols
     newSql = mkCreateTable False def (cols', uniqs, fdef)
     tmpSql = mkCreateTable True def { entityDB = tableTmp } (cols', uniqs, [])
-    dropTmp = "DROP TABLE " <> escape tableTmp
-    dropOld = "DROP TABLE " <> escape table
+    dropTmp = "DROP TABLE " <> escapeE tableTmp
+    dropOld = "DROP TABLE " <> escapeE table
     copyToTemp common = T.concat
         [ "INSERT INTO "
-        , escape tableTmp
+        , escapeE tableTmp
         , "("
-        , T.intercalate "," $ map escape common
+        , T.intercalate "," $ map escapeF common
         , ") SELECT "
-        , T.intercalate "," $ map escape common
+        , T.intercalate "," $ map escapeF common
         , " FROM "
-        , escape table
+        , escapeE table
         ]
     copyToFinal newCols = T.concat
         [ "INSERT INTO "
-        , escape table
+        , escapeE table
         , " SELECT "
-        , T.intercalate "," $ map escape newCols
+        , T.intercalate "," $ map escapeF newCols
         , " FROM "
-        , escape tableTmp
+        , escapeE tableTmp
         ]
 
 mkCreateTable :: Bool -> EntityDef -> ([Column], [UniqueDef], [ForeignDef]) -> Text
@@ -540,7 +544,7 @@ mkCreateTable isTemp entity (cols, uniqs, fdefs) =
         [ "CREATE"
         , if isTemp then " TEMP" else ""
         , " TABLE "
-        , escape $ entityDB entity
+        , escapeE $ entityDB entity
         , "("
         ]
 
@@ -555,12 +559,12 @@ mkCreateTable isTemp entity (cols, uniqs, fdefs) =
             [ T.drop 1 $ T.concat $ map (sqlColumn isTemp) cols
             , ", PRIMARY KEY "
             , "("
-            , T.intercalate "," $ map (escape . fieldDB) $ compositeFields pdef
+            , T.intercalate "," $ map (escapeF . fieldDB) $ compositeFields pdef
             , ")"
             ]
 
         Nothing ->
-            [ escape $ fieldDB (entityId entity)
+            [ escapeF $ fieldDB (entityId entity)
             , " "
             , showSqlType $ fieldSqlType $ entityId entity
             , " PRIMARY KEY"
@@ -583,7 +587,7 @@ mayGenerated gen = case gen of
 sqlColumn :: Bool -> Column -> Text
 sqlColumn noRef (Column name isNull typ def gen _cn _maxLen ref) = T.concat
     [ ","
-    , escape name
+    , escapeF name
     , " "
     , showSqlType typ
     , if isNull then " NULL" else " NOT NULL"
@@ -592,7 +596,7 @@ sqlColumn noRef (Column name isNull typ def gen _cn _maxLen ref) = T.concat
     , case ref of
         Nothing -> ""
         Just ColumnReference {crTableName=table, crFieldCascade=cascadeOpts} ->
-          if noRef then "" else " REFERENCES " <> escape table
+          if noRef then "" else " REFERENCES " <> escapeE table
             <> onDelete cascadeOpts <> onUpdate cascadeOpts
     ]
   where
@@ -602,13 +606,13 @@ sqlColumn noRef (Column name isNull typ def gen _cn _maxLen ref) = T.concat
 sqlForeign :: ForeignDef -> Text
 sqlForeign fdef = T.concat $
     [ ", CONSTRAINT "
-    , escape $ foreignConstraintNameDBName fdef
+    , escapeC $ foreignConstraintNameDBName fdef
     , " FOREIGN KEY("
-    , T.intercalate "," $ map (escape . snd. fst) $ foreignFields fdef
+    , T.intercalate "," $ map (escapeF . snd. fst) $ foreignFields fdef
     , ") REFERENCES "
-    , escape $ foreignRefTableDBName fdef
+    , escapeE $ foreignRefTableDBName fdef
     , "("
-    , T.intercalate "," $ map (escape . snd . snd) $ foreignFields fdef
+    , T.intercalate "," $ map (escapeF . snd . snd) $ foreignFields fdef
     , ")"
     ] ++ onDelete ++ onUpdate
   where
@@ -628,14 +632,23 @@ sqlForeign fdef = T.concat $
 sqlUnique :: UniqueDef -> Text
 sqlUnique (UniqueDef _ cname cols _) = T.concat
     [ ",CONSTRAINT "
-    , escape cname
+    , escapeC cname
     , " UNIQUE ("
-    , T.intercalate "," $ map (escape . snd) cols
+    , T.intercalate "," $ map (escapeF . snd) cols
     , ")"
     ]
 
-escape :: DBName -> Text
-escape (DBName s) =
+escapeC :: ConstraintNameDB -> Text
+escapeC = escapeWith escape
+
+escapeE :: EntityNameDB -> Text
+escapeE = escapeWith escape
+
+escapeF :: FieldNameDB -> Text
+escapeF = escapeWith escape
+
+escape :: Text -> Text
+escape s =
     T.concat [q, T.concatMap go s, q]
   where
     q = T.singleton '"'
@@ -646,21 +659,21 @@ putManySql :: EntityDef -> Int -> Text
 putManySql ent n = putManySql' conflictColumns fields ent n
   where
     fields = entityFields ent
-    conflictColumns = concatMap (map (escape . snd) . uniqueFields) (entityUniques ent)
+    conflictColumns = concatMap (map (escapeF . snd) . uniqueFields) (entityUniques ent)
 
 repsertManySql :: EntityDef -> Int -> Text
 repsertManySql ent n = putManySql' conflictColumns fields ent n
   where
     fields = keyAndEntityFields ent
-    conflictColumns = escape . fieldDB <$> entityKeyFields ent
+    conflictColumns = escapeF . fieldDB <$> entityKeyFields ent
 
 putManySql' :: [Text] -> [FieldDef] -> EntityDef -> Int -> Text
 putManySql' conflictColumns fields ent n = q
   where
-    fieldDbToText = escape . fieldDB
+    fieldDbToText = escapeF . fieldDB
     mkAssignment f = T.concat [f, "=EXCLUDED.", f]
 
-    table = escape . entityDB $ ent
+    table = escapeE . entityDB $ ent
     columns = Util.commaSeparated $ map fieldDbToText fields
     placeholders = map (const "?") fields
     updates = map (mkAssignment . fieldDbToText) fields
@@ -756,12 +769,12 @@ instance FromJSON SqliteConnectionInfo where
 --
 -- @since 2.10.2
 withRawSqliteConnInfo
-    :: (MonadUnliftIO m, MonadLogger m)
+    :: (MonadUnliftIO m, MonadLoggerIO m)
     => SqliteConnectionInfo
     -> (RawSqlite SqlBackend -> m a)
     -> m a
 withRawSqliteConnInfo connInfo f = do
-    logFunc <- askLogFunc
+    logFunc <- askLoggerIO
     withRunInIO $ \run -> E.bracket (openBackend logFunc) closeBackend $ run . f
   where
     openBackend = openWith RawSqlite connInfo
@@ -776,7 +789,7 @@ withRawSqliteConnInfo connInfo f = do
 --
 -- @since 2.10.6
 createRawSqlitePoolFromInfo
-    :: (MonadLogger m, MonadUnliftIO m)
+    :: (MonadLoggerIO m, MonadUnliftIO m)
     => SqliteConnectionInfo
     -> (RawSqlite SqlBackend -> m ())
     -- ^ An action that is run whenever a new `RawSqlite` connection is
@@ -797,7 +810,7 @@ createRawSqlitePoolFromInfo connInfo f n = do
 --
 -- @since 2.10.6
 createRawSqlitePoolFromInfo_
-    :: (MonadLogger m, MonadUnliftIO m)
+    :: (MonadLoggerIO m, MonadUnliftIO m)
     => SqliteConnectionInfo -> Int -> m (Pool (RawSqlite SqlBackend))
 createRawSqlitePoolFromInfo_ connInfo =
   createRawSqlitePoolFromInfo connInfo (const (return ()))
@@ -806,7 +819,7 @@ createRawSqlitePoolFromInfo_ connInfo =
 --
 -- @since 2.10.6
 withRawSqlitePoolInfo
-    :: (MonadUnliftIO m, MonadLogger m)
+    :: (MonadUnliftIO m, MonadLoggerIO m)
     => SqliteConnectionInfo
     -> (RawSqlite SqlBackend -> m ())
     -> Int -- ^ number of connections to open
@@ -824,7 +837,7 @@ withRawSqlitePoolInfo connInfo f n work = do
 --
 -- @since 2.10.6
 withRawSqlitePoolInfo_
-    :: (MonadUnliftIO m, MonadLogger m)
+    :: (MonadUnliftIO m, MonadLoggerIO m)
     => SqliteConnectionInfo
     -> Int -- ^ number of connections to open
     -> (Pool (RawSqlite SqlBackend) -> m a)